Delete old helm releases regardless of status (#14)

+++ prune-releases.sh
@@ -136,9 +136,9 @@ while read release_line ; do
         [ -z "$dry_run" ] && helm delete --namespace $release_namespace $release_name
 
         # Delete the namespace if there are no other helm releases in it
-        if [ "$(helm list --namespace $release_namespace --output json | jq ". | length")" -eq 0 ]; then
+        if [ "$(helm list -a --namespace $release_namespace --output json | jq ". | length")" -eq 0 ]; then
             [ -z "$dry_run" ] && kubectl delete ns $release_namespace
         fi
     fi
-done < <(helm ls --all-namespaces --date --reverse)
+done < <(helm ls -a --all-namespaces --date --reverse)
 [ $counter_delete -gt 0 ] || echo "No stale Helm charts found."
